+++ index.html
@@ -91,6 +91,7 @@
         </div>
     </div>
     <button type="button" class="logout-button">Log In</button>
+    <script src="variables.js"></script>
     <script src="main.js"></script>
 </body>
 </html>
