+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "express": "^4.16.3",
     "jsonwebtoken": "^8.3.0",
+    "node-fetch": "^2.2.0",
     "pg-promise": "^8.4.6"
   }
 }
