+++ server.js
@@ -1,10 +1,11 @@
 const jwt = require('jsonwebtoken');
 const express = require('express');
 const pg = require('pg-promise')();
-const {SIGNATURE, name } = require('./variables');
+const {SIGNATURE, name, recipeKey } = require('./variables');
 const dbConfig = name;
 const db = pg(dbConfig);
 const fs = require('fs');
+const fetch = require('node-fetch');
 
 //Create token
 let createToken = user => {
@@ -66,6 +67,64 @@ let checkToken = async (req, res, next) => {
         res.send('Woops! you do not have a token!');
     }
 };
+
+let getRecipesfromIngreds = (foodArr) => {
+    let prefixUrl =
+      "https://spoonacular-recipe-food-nutrition-v1.p.mashape.com/recipes/findByIngredients?fillIngredients=false&ingredients=";
+    let suffixUrl = "&limitLicense=false&number=5&ranking=2"
+    let ingreRootUrl = foodArr.join("%2C");
+    fetch(prefixUrl + ingreRootUrl + suffixUrl, {
+        method: "GET",
+        headers: {
+            "X-Mashape-Key": recipeKey,
+            Accept: 'application/json'
+        }
+    })
+        .then(function (result) {
+            let promiseRecipes = result.json();
+            return promiseRecipes;
+        })
+        .then(function(recipeObjArr) {
+            // console.log(recipeObjArr);
+            let recipeArrIds = recipeObjArr.map(recipes => recipes.id);
+            getRecipeInfo(recipeArrIds);
+            })
+        .catch(function (err) {
+            console.log(err);
+        });
+}
+
+
+let getRecipeInfo = function(recipeArrIds) {
+    let prefixUrl = 'https://spoonacular-recipe-food-nutrition-v1.p.mashape.com/recipes/informationBulk?ids=';
+    let suffixUrl = '&includeNutrition=false';
+    let ingreRootUrl = recipeArrIds.join("%2C");
+    let newRecipes = [];
+    fetch(prefixUrl + ingreRootUrl + suffixUrl, {
+        method: "GET",
+        headers: {
+            "X-Mashape-Key": recipeKey,
+            Accept: 'application/json'
+        }
+    })
+        .then(function (result) {
+            let promiseRecipes = result.json();
+            return promiseRecipes;
+        })
+        .then(function(recipeObjArr) {
+            console.log(recipeObjArr);
+            let newValues = recipeObjArr.map(recipe => {
+                newRecipes.push([recipe.title, recipe.spoonacularSourceUrl, recipe.image]);
+            })
+            console.log(newRecipes);
+            })
+        .catch(function (err) {
+            console.log(err);
+        });
+}
+
+console.log(getRecipesfromIngreds(['sugar', 'apple', 'flour']));
+
 
 let postUserSignupInformation = (req, res) => {
     readBody(req, (body) => {
